requirements.txt
moviepy==1.0.1
numpy
Pillow